+++ pkg/cloud/cache/cache.go
@@ -16,11 +16,13 @@ const (
 	zoneRecordsPrefix           = "zoneRecords"
 	zoneIDPrefix                = "zoneID"
 	nameserverRecordsPrefix     = "nameserverRecords"
+	clusterGatewayRecordsPrefix = "gatewayRecords"
 
 	ClusterIngressRecords = 1
 	ZoneRecords           = 2
 	ZoneID                = 3
 	NameserverRecords     = 4
+	ClusterGatewayRecords = 5
 
 	unknownCacheIDError = "unknown cache identifier"
 )
@@ -80,6 +82,8 @@ func GetDNSCacheRecord(recordID int, keySuffix string) ([]byte, error) {
 		return DNSOperatorCache.Get(fmt.Sprintf("%s-%s", zoneIDPrefix, keySuffix))
 	case NameserverRecords:
 		return DNSOperatorCache.Get(fmt.Sprintf("%s-%s", nameserverRecordsPrefix, keySuffix))
+	case ClusterGatewayRecords:
+		return DNSOperatorCache.Get(fmt.Sprintf("%s-%s", clusterGatewayRecordsPrefix, keySuffix))
 	default:
 		return nil, errors.New(unknownCacheIDError)
 	}
@@ -97,6 +101,8 @@ func SetDNSCacheRecord(recordID int, keySuffix string, data []byte) error {
 		return DNSOperatorCache.Set(fmt.Sprintf("%s-%s", zoneIDPrefix, keySuffix), data)
 	case NameserverRecords:
 		return DNSOperatorCache.Set(fmt.Sprintf("%s-%s", nameserverRecordsPrefix, keySuffix), data)
+	case ClusterGatewayRecords:
+		return DNSOperatorCache.Set(fmt.Sprintf("%s-%s", clusterGatewayRecordsPrefix, keySuffix), data)
 	default:
 		return errors.New(unknownCacheIDError)
 	}
@@ -114,6 +120,8 @@ func DeleteDNSCacheRecord(recordID int, keySuffix string) error {
 		return DNSOperatorCache.Delete(fmt.Sprintf("%s-%s", zoneIDPrefix, keySuffix))
 	case NameserverRecords:
 		return DNSOperatorCache.Delete(fmt.Sprintf("%s-%s", nameserverRecordsPrefix, keySuffix))
+	case ClusterGatewayRecords:
+		return DNSOperatorCache.Delete(fmt.Sprintf("%s-%s", clusterGatewayRecordsPrefix, keySuffix))
 	default:
 		return errors.New(unknownCacheIDError)
 	}