+++ pkg/key/key.go
@@ -3,4 +3,6 @@ package key
 const (
 	DNSFinalizerNameOld = "dns-operator-openstack.finalizers.giantswarm.io"
 	DNSFinalizerNameNew = "dns-operator-route53.finalizers.giantswarm.io"
+
+	AnnotationWildcardCNAMETarget = "network.giantswarm.io/wildcard-cname-target"
 )